MultimodalDamagePrediction/requirements.txt
torch>=2.0.0
torchvision
timm
numpy
Pillow
pyyaml
streamlit
opencv-python
tqdm
